Changed client to use request struct
the shell commands are parsed in the docmd() function.

This function populate the global reqest struct and call the doRPCCmd()
function for RPC communication.

This allows us to capture the request and in the future filter requests
(i.e not actually send the reqests if the client's game state already
knows the response)

Untested  ... neeed to get basic server working first.