Added Skeleton for Server
Now the server has code for responding to the different RPC types...
next step is to connect client/server and then to implement game logic
ontop of this skeleton.

It might be easier to compact all the num* queries into one query where
we ask for cell count with a specified cell type floor/wall/jail/home
etc.