dynamic home flag tracking
Whenever a player enters or exists a home cell with a flag, we now
update the homes new 'flags' attribute with the number of flags. in the
cell.  This allows us to compute the game state dynamically on the fly.

update the server locks test to ensure changes occur atomically.