implement and test update_wait/signal
Implemented the update wait and signal methods for server use.

The test spawns 1000 threads, each with a timestamp between 0-1000.
The thread is then forced to wait until it is its turn. When it is its
turn the thread will save the realtime to a slot in an array
corresponding to the expected order.  Then the thread broadcasts, waking
all threads tell them to check if any of them are next.